Don't pass credentials to the CSI driver via environment variables
