+++ assets/csi_cloud_config.ini
@@ -2,6 +2,8 @@
 
 [Global]
 cluster-id = "${CLUSTER_ID}"
+user = "${USER}"
+password = "${PASSWORD}"
 
 [VirtualCenter "${VCENTER}"]
 insecure-flag = "true"
