+++ pkg/operator/vspherecontroller/driver_starter.go
@@ -6,12 +6,9 @@ import (
 	"fmt"
 	"os"
 
-	infralister "github.com/openshift/client-go/config/listers/config/v1"
 	"github.com/openshift/library-go/pkg/operator/resource/resourcehash"
 	"github.com/openshift/vmware-vsphere-csi-driver-operator/pkg/operator/utils"
-	"gopkg.in/gcfg.v1"
 
-	ocpv1 "github.com/openshift/api/config/v1"
 	operatorapi "github.com/openshift/api/operator/v1"
 	"github.com/openshift/library-go/pkg/controller/factory"
 	"github.com/openshift/library-go/pkg/operator/csi/csicontrollerset"
@@ -24,9 +21,6 @@ import (
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	corev1informers "k8s.io/client-go/informers/core/v1"
-	corelister "k8s.io/client-go/listers/core/v1"
-	"k8s.io/klog/v2"
-	"k8s.io/legacy-cloud-providers/vsphere"
 )
 
 func (c *VSphereController) createCSIDriver() {
@@ -113,7 +107,6 @@ func (c *VSphereController) createCSIDriver() {
 			c.apiClients.ConfigMapInformer.Informer(),
 			c.apiClients.NodeInformer.Informer(),
 		},
-		WithVSphereCredentials(defaultNamespace, cloudCredSecretName, c.infraLister, c.configMapLister, c.apiClients.SecretInformer),
 		WithSyncerImageHook("vsphere-syncer"),
 		WithLogLevelDeploymentHook(),
 		c.topologyHook,
@@ -184,91 +177,6 @@ func (c *VSphereController) topologyHook(opSpec *operatorapi.OperatorSpec, deplo
 		deployment.Spec.Template.Spec.Containers = containers
 	}
 	return nil
-}
-
-func WithVSphereCredentials(
-	namespace string,
-	secretName string,
-	infraLister infralister.InfrastructureLister,
-	configMapLister corelister.ConfigMapLister,
-	secretInformer corev1informers.SecretInformer,
-) deploymentcontroller.DeploymentHookFunc {
-	return func(opSpec *operatorapi.OperatorSpec, deployment *appsv1.Deployment) error {
-		secret, err := secretInformer.Lister().Secrets(namespace).Get(secretName)
-		if err != nil {
-			return err
-		}
-
-		infra, err := infraLister.Get(infraGlobalName)
-		if err != nil {
-			return err
-		}
-
-		vCenterName, err := getvCenterName(infra, configMapLister)
-		if err != nil {
-			return err
-		}
-
-		// CCO generates a secret that contains dynamic keys, for example:
-		// oc get secret/vmware-vsphere-cloud-credentials -o json | jq .data
-		// {
-		//   "vcenter.xyz.vmwarevmc.com.password": "***",
-		//   "vcenter.xyz.vmwarevmc.com.username": "***"
-		// }
-		// So we need to figure those keys out
-		var usernameKey, passwordKey string
-
-		if len(secret.Data) > 2 {
-			klog.Warningf("CSI driver can only connect to one vcenter, more than 1 set of credentials found for CSI driver")
-		}
-
-		usernameKey = vCenterName + ".username"
-		passwordKey = vCenterName + ".password"
-
-		if usernameKey == "" || passwordKey == "" {
-			return fmt.Errorf("could not find vSphere credentials in secret %s/%s", secret.Namespace, secret.Name)
-		}
-
-		// Add to csi-driver and vsphere-syncer containers the vSphere credentials, as env vars.
-		containers := deployment.Spec.Template.Spec.Containers
-		for i := range containers {
-			if containers[i].Name != "csi-driver" && containers[i].Name != "vsphere-syncer" {
-				continue
-			}
-			containers[i].Env = append(
-				containers[i].Env,
-				newSecretEnvVar(secretName, "VSPHERE_USER", usernameKey),
-				newSecretEnvVar(secretName, "VSPHERE_PASSWORD", passwordKey),
-			)
-		}
-		deployment.Spec.Template.Spec.Containers = containers
-		return nil
-	}
-}
-
-func getvCenterName(infra *ocpv1.Infrastructure, configmapLister corelister.ConfigMapLister) (string, error) {
-	// This change can only be used in >=4.13 versions of OCP
-	vSphereInfraConfig := infra.Spec.PlatformSpec.VSphere
-	if vSphereInfraConfig != nil && len(vSphereInfraConfig.VCenters) > 0 {
-		return vSphereInfraConfig.VCenters[0].Server, nil
-	}
-
-	cloudConfig := infra.Spec.CloudConfig
-	cloudConfigMap, err := configmapLister.ConfigMaps(cloudConfigNamespace).Get(cloudConfig.Name)
-	if err != nil {
-		return "", fmt.Errorf("failed to get cloud config: %v", err)
-	}
-
-	cfgString, ok := cloudConfigMap.Data[infra.Spec.CloudConfig.Key]
-	if !ok {
-		return "", fmt.Errorf("cloud config %s/%s does not contain key %q", cloudConfigNamespace, cloudConfig.Name, cloudConfig.Key)
-	}
-	cfg := new(vsphere.VSphereConfig)
-	err = gcfg.ReadStringInto(cfg, cfgString)
-	if err != nil {
-		return "", err
-	}
-	return cfg.Workspace.VCenterIP, nil
 }
 
 func WithSyncerImageHook(containerName string) deploymentcontroller.DeploymentHookFunc {
