+++ pkg/operator/vspherecontroller/vspherecontroller.go
@@ -29,6 +29,7 @@ import (
 	"github.com/openshift/vmware-vsphere-csi-driver-operator/pkg/operator/utils"
 	"github.com/openshift/vmware-vsphere-csi-driver-operator/pkg/operator/vclib"
 	"github.com/openshift/vmware-vsphere-csi-driver-operator/pkg/operator/vspherecontroller/checks"
+	corev1informers "k8s.io/client-go/informers/core/v1"
 	"k8s.io/client-go/kubernetes"
 	corelister "k8s.io/client-go/listers/core/v1"
 	storagelister "k8s.io/client-go/listers/storage/v1"
@@ -662,12 +663,18 @@ func (c *VSphereController) applyClusterCSIDriverChange(
 	}
 
 	datacenters := strings.Join(dataCenterNames, ",")
+	user, password, err := getUserAndPassword(defaultNamespace, cloudCredSecretName, infra, c.configMapLister, c.apiClients.SecretInformer)
+	if err != nil {
+		return nil, err
+	}
 
 	for pattern, value := range map[string]string{
 		"${CLUSTER_ID}":              infra.Status.InfrastructureName,
 		"${VCENTER}":                 sourceCFG.Workspace.VCenterIP,
 		"${DATACENTERS}":             datacenters,
 		"${MIGRATION_DATASTORE_URL}": datastoreURL,
+		"${USER}":                    user,
+		"${PASSWORD}":                password,
 	} {
 		csiConfigString = strings.ReplaceAll(csiConfigString, pattern, value)
 	}
@@ -716,4 +723,65 @@ func (c *VSphereController) createStorageClassController() storageclasscontrolle
 		c.eventRecorder,
 	)
 	return storageClassController
+}
+
+func getUserAndPassword(namespace string, secretName string, infra *ocpv1.Infrastructure, configMapLister corelister.ConfigMapLister, secretInformer corev1informers.SecretInformer,
+) (string, string, error) {
+	secret, err := secretInformer.Lister().Secrets(namespace).Get(secretName)
+	if err != nil {
+		return "", "", err
+	}
+
+	vCenterName, err := getvCenterName(infra, configMapLister)
+	if err != nil {
+		return "", "", err
+	}
+
+	// CCO generates a secret that contains dynamic keys, for example:
+	// oc get secret/vmware-vsphere-cloud-credentials -o json | jq .data
+	// {
+	//   "vcenter.xyz.vmwarevmc.com.password": "***",
+	//   "vcenter.xyz.vmwarevmc.com.username": "***"
+	// }
+	// So we need to figure those keys out
+	var usernameKey, passwordKey string
+
+	if len(secret.Data) > 2 {
+		klog.Warningf("CSI driver can only connect to one vcenter, more than 1 set of credentials found for CSI driver")
+	}
+
+	usernameKey = vCenterName + ".username"
+	passwordKey = vCenterName + ".password"
+
+	if usernameKey == "" || passwordKey == "" {
+		return "", "", fmt.Errorf("could not find vSphere credentials in secret %s/%s", secret.Namespace, secret.Name)
+	}
+
+	// Get user and pass from CCO secret
+	return string(secret.Data[usernameKey]), string(secret.Data[passwordKey]), nil
+}
+
+func getvCenterName(infra *ocpv1.Infrastructure, configmapLister corelister.ConfigMapLister) (string, error) {
+	// This change can only be used in >=4.13 versions of OCP
+	vSphereInfraConfig := infra.Spec.PlatformSpec.VSphere
+	if vSphereInfraConfig != nil && len(vSphereInfraConfig.VCenters) > 0 {
+		return vSphereInfraConfig.VCenters[0].Server, nil
+	}
+
+	cloudConfig := infra.Spec.CloudConfig
+	cloudConfigMap, err := configmapLister.ConfigMaps(cloudConfigNamespace).Get(cloudConfig.Name)
+	if err != nil {
+		return "", fmt.Errorf("failed to get cloud config: %v", err)
+	}
+
+	cfgString, ok := cloudConfigMap.Data[infra.Spec.CloudConfig.Key]
+	if !ok {
+		return "", fmt.Errorf("cloud config %s/%s does not contain key %q", cloudConfigNamespace, cloudConfig.Name, cloudConfig.Key)
+	}
+	cfg := new(vsphere.VSphereConfig)
+	err = gcfg.ReadStringInto(cfg, cfgString)
+	if err != nil {
+		return "", err
+	}
+	return cfg.Workspace.VCenterIP, nil
 }
